--- 2016Roobut/src/camera/CameraMove.java
@@ -1,8 +1,0 @@
-package camera;
-
-public class CameraMove {
-//Million put code here.
-//That's not how you spell my name!
-
-	//Code for rotating Camera
-}
